# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenLicensd API
  description: |
    HTTP API for OpenLicensd — an open source license server for creating and validating license keys.

    Admin endpoints require a session cookie obtained from `POST /api/v1/auth/login` or OIDC SSO.
    Unsafe methods also require the `X-CSRF-Token` header matching the `openlicensd_csrf` cookie.
    Public endpoints (`/validate`, `/registry-credentials`, `/auth/login`, OIDC login/callback) do not require authentication.
    Unauthenticated endpoints are rate limited per client IP; configure `OPENLICENSD_TRUSTED_PROXIES` behind a reverse proxy.

    All responses include browser security headers: `Content-Security-Policy`, `X-Frame-Options`,
    and `X-Content-Type-Options`. When `OPENLICENSD_COOKIE_SECURE=true`, responses also include
    `Strict-Transport-Security`.

    The `/api/v1/registry-credentials` endpoint is only available when
    `OPENLICENSD_HARBOR_ENABLED=true` on the server.

    ## Role-based access control

    Authenticated endpoints enforce one of three roles. Insufficient role returns `403` with
    `{"error":"forbidden"}`.

    | Role | Permissions |
    |------|-------------|
    | `admin` | Full access including user management |
    | `operator` | Create, update, and delete licenses, products, and policies |
    | `viewer` | Read-only access to licenses, products, and policies |

    Any authenticated user (any role) may call `/api/v1/auth/me`, `/api/v1/auth/logout`, and
...
  license:
...
servers:
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Admin login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Local login is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
      security: []
...
  /api/v1/auth/oidc/login:
    get:
      tags: [Auth]
      summary: Start OIDC login
      description: Redirects to the configured identity provider. Only available when `OPENLICENSD_OIDC_ENABLED=true`.
      operationId: oidcLogin
      parameters:
        - name: return_to
          in: query
          required: false
          schema:
            type: string
          description: Relative path to redirect after successful login (default `/licenses`)
      responses:
        "302":
          description: Redirect to identity provider
        "404":
          description: OIDC is not enabled
        "429":
          $ref: "#/components/responses/TooManyRequests"
      security: []
  /api/v1/auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC callback
      description: Handles the authorization code callback from the identity provider. Sets session cookies and redirects into the app.
      operationId: oidcCallback
      responses:
        "302":
          description: Redirect to app or login page on failure
        "404":
          description: OIDC is not enabled
        "429":
          $ref: "#/components/responses/TooManyRequests"
      security: []
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: |
        Returns the currently authenticated user.

        **Required role:** any authenticated user (`admin`, `operator`, or `viewer`).
      operationId: getCurrentUser
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Log out
      description: |
        Revokes the current session and clears session cookies.

        **Required role:** any authenticated user (`admin`, `operator`, or `viewer`).
      operationId: logout
      security:
        - sessionAuth: []
      responses:
        "204":
          description: Logged out successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/password:
    post:
      tags: [Auth]
      summary: Change own password
      description: |
        Changes the password for the currently authenticated user. Requires the current
        password. Only available for accounts with a local password (`has_password: true`).
        On success, all other sessions for the user are revoked; the current session
        remains valid.

        **Required role:** any authenticated user (`admin`, `operator`, or `viewer`).
      operationId: changeOwnPassword
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed successfully
        "400":
          description: Invalid request (wrong current password, password too short, or account has no local password)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      description: |
        **Required role:** `admin`.
      operationId: listUsers
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/UserSort"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        "400":
          description: Invalid list parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Users]
      summary: Create a user
      description: |
        Creates a local user with a bcrypt-hashed password.

        **Required role:** `admin`.
      operationId: createUser
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request body (missing email, name, or password; invalid role; or password too short)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserId"

    patch:
      tags: [Users]
      summary: Update a user
      description: |
        Updates a user's email, name, and role.

        **Required role:** `admin`.
      operationId: updateUser
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user ID or request body (missing email, name, or invalid role)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Users]
      summary: Delete a user
      description: |
        Permanently deletes a user. Cannot delete your own account.

        **Required role:** `admin`.
      operationId: deleteUser
      security:
        - sessionAuth: []
      responses:
        "204":
          description: User deleted
        "400":
          description: Invalid user ID or cannot delete your own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/users/{id}/password:
    parameters:
      - $ref: "#/components/parameters/UserId"

    patch:
      tags: [Users]
      summary: Set user password
      description: |
        Sets or resets the password for a user (admin action). Does not require the current
        password. Minimum length is 8 characters. Store failures return `404`.

        **Required role:** `admin`.
      operationId: setUserPassword
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserPasswordRequest"
      responses:
        "204":
          description: Password set successfully
        "400":
          description: Invalid user ID, missing password, or password too short
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{id}/disable:
    parameters:
      - $ref: "#/components/parameters/UserId"

    patch:
      tags: [Users]
      summary: Disable a user
      description: |
        Disables a user account and revokes all of their sessions. Cannot disable your own account.

        **Required role:** `admin`.
      operationId: disableUser
      security:
        - sessionAuth: []
      responses:
        "200":
          description: User disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user ID or cannot disable your own account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/users/{id}/enable:
    parameters:
      - $ref: "#/components/parameters/UserId"

    patch:
      tags: [Users]
      summary: Enable a user
      description: |
        Re-enables a disabled user and clears failed-login lockout counters.

        **Required role:** `admin`.
      operationId: enableUser
      security:
        - sessionAuth: []
      responses:
        "200":
          description: User enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products
      description: |
        **Required role:** `viewer`, `operator`, or `admin`.
      operationId: listProducts
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/ProductSort"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"
        "400":
          description: Invalid list parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Products]
      summary: Create a product
      description: |
        **Required role:** `operator` or `admin`.
      operationId: createProduct
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/products/{id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"

    patch:
      tags: [Products]
      summary: Update a product
      description: |
        **Required role:** `operator` or `admin`.
      operationId: updateProduct
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product ID or request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Products]
      summary: Delete a product
      description: |
        **Required role:** `operator` or `admin`.
      operationId: deleteProduct
      security:
        - sessionAuth: []
      responses:
        "204":
          description: Product deleted
        "400":
          description: Invalid product ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/policies:
    get:
      tags: [Policies]
      summary: List policies
      description: |
        **Required role:** `viewer`, `operator`, or `admin`.
      operationId: listPolicies
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/PolicySort"
        - $ref: "#/components/parameters/Order"
        - name: product_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter policies by product
      responses:
        "200":
          description: Paginated list of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPolicies"
        "400":
          description: Invalid list parameters or product_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Policies]
      summary: Create a policy
      description: |
        **Required role:** `operator` or `admin`.
      operationId: createPolicy
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/policies/{id}:
    parameters:
      - $ref: "#/components/parameters/PolicyId"

    patch:
      tags: [Policies]
      summary: Update a policy
      description: |
        **Required role:** `operator` or `admin`.
      operationId: updatePolicy
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePolicyRequest"
      responses:
        "200":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          description: Invalid policy ID or request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Policies]
      summary: Delete a policy
      description: |
        **Required role:** `operator` or `admin`.
      operationId: deletePolicy
      security:
        - sessionAuth: []
      responses:
        "204":
          description: Policy deleted
        "400":
          description: Invalid policy ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/licenses/stats:
    get:
      tags: [Licenses]
      summary: License status counts
      description: |
        **Required role:** `viewer`, `operator`, or `admin`.
      operationId: getLicenseStats
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Aggregate license counts by status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/licenses:
    get:
      tags: [Licenses]
      summary: List licenses
      description: |
        **Required role:** `viewer`, `operator`, or `admin`.
      operationId: listLicenses
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/LicenseSort"
        - $ref: "#/components/parameters/Order"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, expired, revoked]
          description: Filter licenses by status
        - name: product_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter licenses by product
        - name: policy_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter licenses by policy
      responses:
        "200":
          description: Paginated list of licenses (raw keys are never returned)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLicenses"
        "400":
          description: Invalid list parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Licenses]
      summary: Create a license
      description: |
        Generates a new license key and returns it once in the response.
        Only the key prefix is stored for display; the full key cannot be retrieved later.

        **Required role:** `operator` or `admin`.
      operationId: createLicense
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLicenseRequest"
      responses:
        "201":
          description: License created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseCreated"
        "400":
          description: Invalid request body, missing label, or policy does not belong to product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/licenses/{id}/revoke:
    parameters:
      - $ref: "#/components/parameters/LicenseId"
    patch:
      tags: [Licenses]
      summary: Revoke a license
      description: |
        Revoked licenses fail validation immediately.

        **Required role:** `operator` or `admin`.
      operationId: revokeLicense
      security:
        - sessionAuth: []
      responses:
        "200":
          description: License revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "400":
          description: Invalid license ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/licenses/{id}/unrevoke:
    parameters:
      - $ref: "#/components/parameters/LicenseId"
    patch:
      tags: [Licenses]
      summary: Unrevoke a license
      description: |
        Clears the license `revoked` flag so validation succeeds again. This does not create a
        machine activation or set `activated_at` (those are separate: machine seats via
        validation fingerprints, and first-validation policy via `activated_at`).

        **Required role:** `operator` or `admin`.
      operationId: unrevokeLicense
      security:
        - sessionAuth: []
      responses:
        "200":
          description: License unrevoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "400":
          description: Invalid license ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: openlicensd_session
      description: |
        Session cookie obtained from `POST /api/v1/auth/login` or OIDC SSO.
        Unsafe methods (POST, PATCH, DELETE) also require the `X-CSRF-Token` header
        set to the value of the `openlicensd_csrf` cookie.
...
  responses:
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until the client may retry
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: rate limit exceeded
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
...